requirements.txt
flask
flask-cors
requests
beautifulsoup4
lxml
groq
python-dotenv
urllib3
certifi